Move RepositoryMetrics to Records namespace #20
