+++ RepoMan/RepoMan/Analysis/RepositoryAnalyzer.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Newtonsoft.Json;
 using RepoMan.Records;
 
 namespace RepoMan.Analysis
@@ -39,31 +38,5 @@ namespace RepoMan.Analysis
             };
             return repoMetrics;
         }
-    }
-    
-    public class RepositoryMetrics
-    {
-        public DateTimeOffset Timestamp { get; set; }
-        
-        public HashSet<int> PullRequests { get; set; }
-        public int PullRequestCount => PullRequests.Count;
-        
-        public int MedianCommentCountPerPullRequest { get; set; }
-        public int MedianWordsPerComment { get; set; }
-        
-        [JsonConverter(typeof(TruncatingDoubleConverter))]
-        public double CommentCountPopulationVariance { get; set; }
-        
-        [JsonConverter(typeof(TruncatingDoubleConverter))]
-        public double CommentCountPopulationStdDeviation => Math.Sqrt(CommentCountPopulationVariance); 
-        
-        [JsonConverter(typeof(TruncatingDoubleConverter))]
-        public double CommentWordCountVariance { get; set; }
-        
-        [JsonConverter(typeof(TruncatingDoubleConverter))]
-        public double CommentWordCountStdDeviation => Math.Sqrt(CommentWordCountVariance);
-        
-        public int MedianSecondsToPullRequestClosure { get; set; }
-        public int MedianBusinessDaysToPullRequestClosure { get; set; }
     }
 }
