+++ RepoMan/RepoMan/IO/FilesystemDataProvider.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
-using RepoMan.Analysis;
 using RepoMan.Records;
 
 namespace RepoMan.IO
